+++ meterparser/src/app/mqtt.py
@@ -25,6 +25,14 @@ class Mqtt:
         self._mqtt_client.on_disconnect = self.mqtt_disconnected
 
         self.cameras: list = list()
+        self.device_id = slugify((os.environ["HOSTNAME"] if "HOSTNAME" in os.environ else os.environ["COMPUTERNAME"]).lower())
+        self.device = {
+                "identifiers": self.device_id,
+                "manufacturer": "Meter Parser",
+                "model": "Meter Parser Add-On",
+                "name": "Meter Parser at %s" % self.device_id,
+                "sw_version": "1.0.0.6" # TODO: Get from git build
+            }
     def mqtt_connected(self, client, userdata, flags, rc):
         # spawn camera threads
         from app.camera import Camera
@@ -60,59 +68,50 @@ class Mqtt:
                 _LOGGER.error("Could not connect to mqtt: %s. Retry in 5 secs." % e)
                 time.sleep(5)
     def mqtt_sensor_discovery(self, entity_id: str, name: str, device_class: str, unit_of_measurement: str):
-        device_id = os.environ["HOSTNAME"] if "HOSTNAME" in os.environ else os.environ["COMPUTERNAME"]
-        device = {
-                "identifiers": device_id,
-                "manufacturer": "Meter Parser",
-                "model": "Meter Parser Add-On",
-                "name": "Meter Parser",
-                "sw_version": "1.0.0" # TODO: Get from git build            
-            }
-            
-        topic_sensor = "%s/sensor/%s/config" % (discovery_prefix, entity_id)
+           
+        topic_sensor = "%s/sensor/%s/%s/config" % (discovery_prefix, self.device_id, entity_id)
         icon = "mdi:water"
         if device_class == "energy":
             icon = "mdi:flash"
         elif device_class == "gas":
             icon = "mdi:fire"
 
-
         payload_sensor = {
             "name": name, 
             "icon": icon,
             "unit_of_measurement": unit_of_measurement,
             "state_class": "total_increasing",
-            "state_topic": "%s/sensor/%s/state" % (discovery_prefix, entity_id),
-            "availability_topic": "%s/sensor/%s/availability" % (discovery_prefix, entity_id),
-            "json_attributes_topic": "%s/sensor/%s/attributes" % (discovery_prefix, entity_id),
-            "object_id": entity_id,
-            "unique_id": "%s.%s" % (device_id, entity_id),
-            "device": device
+            "state_topic": "%s/sensor/%s/%s/state" % (discovery_prefix, self.device_id, entity_id),
+            "availability_topic": "%s/sensor/%s/%s/availability" % (discovery_prefix, self.device_id, entity_id),
+            "json_attributes_topic": "%s/sensor/%s/%s/attributes" % (discovery_prefix, self.device_id, entity_id),
+            "unique_id": "%s_%s" % (self.device_id, entity_id),
+            "device": self.device
         }
         if device_class != "water":
             payload_sensor["device_class"] = device_class
 
 
-        topic_camera = "%s/camera/%s/config" % (discovery_prefix, entity_id)
+        topic_camera = "%s/camera/%s/%s/config" % (discovery_prefix, self.device_id, entity_id)
         payload_camera = {
             "name": name, 
-            "topic": "%s/camera/%s/state" % (discovery_prefix, entity_id),
-            "availability_topic": "%s/camera/%s/availability" % (discovery_prefix, entity_id),
-            "json_attributes_topic": "%s/camera/%s/attributes" % (discovery_prefix, entity_id),
-            "object_id": entity_id,
-            "unique_id": "%s.%s_cam" % (device_id, entity_id),
-            "device": device
+            "topic": "%s/camera/%s/%s/state" % (discovery_prefix, self.device_id, entity_id),
+            "availability_topic": "%s/camera/%s/%s/availability" % (discovery_prefix, self.device_id, entity_id),
+            "json_attributes_topic": "%s/camera/%s/%s/attributes" % (discovery_prefix, self.device_id, entity_id),
+            "unique_id": "%s_%s_cam" % (self.device_id, entity_id),
+            "device": self.device
         }
 
         self._mqtt_client.publish(topic_sensor, payload=json.dumps(payload_sensor).encode("utf-8"))
         self._mqtt_client.publish(topic_camera, payload=json.dumps(payload_camera).encode("utf-8"))
 
     def mqtt_set_state(self, type:str, entity_id: str, state):
-        topic = "%s/%s/%s/state" % (discovery_prefix, type, entity_id)
+        topic = "%s/%s/%s/%s/state" % (discovery_prefix, type, self.device_id, entity_id)
         self._mqtt_client.publish(topic, payload=state)
+
     def mqtt_set_attributes(self, type:str, entity_id: str, attributes):
-        topic = "%s/%s/%s/attributes" % (discovery_prefix, type, entity_id)
+        topic = "%s/%s/%s/%s/attributes" % (discovery_prefix, type, self.device_id, entity_id)
         self._mqtt_client.publish(topic, payload=json.dumps(attributes))
+
     def mqtt_set_availability(self, type:str, entity_id: str, available: bool):
-        topic = "%s/%s/%s/availability" % (discovery_prefix, type, entity_id)
+        topic = "%s/%s/%s/%s/availability" % (discovery_prefix, type, self.device_id, entity_id)
         self._mqtt_client.publish(topic, payload=("online" if available else "offline"))