+++ meterparser/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [1.0.0.*]
+- Rewrite topic to allow concurrent instances.
+- Send current value to sensor on intervals
+- Fix service type on devices
 - Fix incorrect variable name.
 - Save image to PNG for OCR.
 - Add `stdin` input to list and reset the control value.